MyGPT/backend/app/config.py
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

class Config:
    DATABASE_URL = os.getenv("DATABASE_URL")
    API_KEY = os.getenv("API_KEY")
    VECTOR_DB_URL = os.getenv("VECTOR_DB_URL")

config = Config()
