MyGPT/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
ollama
langchain-community
chromadb
unstructured
# python-magic-bin